[audio] merge modules
