+++ packages/expo-audio/ios/AudioModule.swift
@@ -1,12 +1,16 @@
 import ExpoModulesCore
 import Combine
 
+private let status = "onRecordingStatusUpdate"
 private let statusUpdate = "onPlaybackStatusUpdate"
 
-public class AudioModule: Module {
+public class AudioModule: Module, RecordingResultHandler {
   private var timeTokens = [Int: Any?]()
   private var players = [String: AudioPlayer]()
   private var sessionIsActive = true
+  private lazy var recordingDelegate = {
+    RecordingDelegate(resultHandler: self)
+  }()
   
   // Observers
   private var cancellables = Set<AnyCancellable>()
@@ -37,6 +41,22 @@ public class AudioModule: Module {
       } catch {
         throw AudioStateException(error.localizedDescription)
       }
+    }
+    
+    AsyncFunction("requestRecordingPermissionsAsync") { (promise: Promise) in
+      appContext?.permissions?.askForPermission(
+        usingRequesterClass: AudioRecordingRequester.self,
+        resolve: promise.resolver,
+        reject: promise.legacyRejecter
+      )
+    }
+    
+    AsyncFunction("getRecordingPermissionsAsync") { (promise: Promise) in
+      appContext?.permissions?.getPermissionUsingRequesterClass(
+        AudioRecordingRequester.self,
+        resolve: promise.resolver,
+        reject: promise.legacyRejecter
+      )
     }
     
     OnDestroy {
@@ -159,6 +179,97 @@ public class AudioModule: Module {
           )
         )
       }
+    }
+    
+    Class(AudioRecorder.self) {
+      Constructor { (url: String?) -> AudioRecorder in
+        guard let cachesDir = appContext?.fileSystem?.cachesDirectory, var directory = URL(string: cachesDir) else {
+          throw Exceptions.AppContextLost()
+        }
+        
+        directory.appendPathComponent("Recording")
+        FileSystemUtilities.ensureDirExists(at: directory)
+        if let url {
+          directory.appendingPathComponent(url)
+        }
+        
+        return AudioRecorder(createRecorder(url: directory))
+      }
+      
+      Property("isRecording") { recorder in
+        recorder.pointer.isRecording
+      }
+      
+      Property("currentTime") { recorder in
+        recorder.pointer.currentTime
+      }
+        
+      Function("record") { recorder in
+        try checkPermissions()
+        recorder.pointer.record()
+      }
+      
+      Function("stop") { recorder in
+        try checkPermissions()
+        recorder.pointer.stop()
+      }
+      
+      Function("startRecordingAtTime") { (recorder, seconds: Double) in
+        try checkPermissions()
+        recorder.pointer.record(atTime: TimeInterval(seconds))
+      }
+      
+      Function("recordForDuration") { (recorder, seconds: Double) in
+        try checkPermissions()
+        recorder.pointer.record(forDuration: TimeInterval(seconds))
+      }
+    }
+  }
+  
+  func didFinish(_ recorder: AVAudioRecorder, successfully flag: Bool) {
+    sendEvent(status, [
+      "isFinished": true,
+      "hasError": false,
+      "url": recorder.url
+    ])
+  }
+  
+  func encodeErrorDidOccur(_ recorder: AVAudioRecorder, error: Error?) {
+    sendEvent(status, [
+      "isFinished": true,
+      "hasError": true,
+      "error": error?.localizedDescription,
+      "url": nil
+    ])
+  }
+  
+  private func createRecorder(url: URL?) -> AVAudioRecorder {
+    let recorder = {
+      if let url {
+        do {
+          return try AVAudioRecorder(url: url, settings: [
+            AVFormatIDKey: kAudioFormatLinearPCM,
+            AVSampleRateKey: 8.0,
+            AVNumberOfChannelsKey: 64,
+            AVAudioBitRateStrategy_Variable: AVEncoderAudioQualityKey,
+          ])
+        } catch {
+          return AVAudioRecorder()
+        }
+      }
+      return AVAudioRecorder()
+    }()
+    
+    recorder.delegate = recordingDelegate
+    return recorder
+  }
+  
+  private func checkPermissions() throws {
+    switch AVAudioSession.sharedInstance().recordPermission {
+    case .denied, .undetermined:
+      throw AudioPermissionsException()
+    default:
+      break
     }
   }
